+++ src/RemoteStorageDetails/RemoteStorageDetails.js
@@ -1,8 +1,9 @@
 import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import { useIntl } from 'react-intl';
 import {
   useHistory,
+  useParams,
 } from 'react-router-dom';
 
 import {
@@ -15,6 +16,8 @@ import {
   MenuSection,
   Row,
   KeyValue,
+  Icon,
+  Button,
 } from '@folio/stripes/components';
 import {
   ViewMetaData,
@@ -32,8 +35,11 @@ const RemoteStorageDetails = ({
   defaultWidth,
   storage,
   isLoading,
+  onRemovestorage,
 }) => {
   const history = useHistory();
+  const { id } = useParams();
+  const intl = useIntl();
 
   const [expandAll, sections, toggleSection] = useAccordionToggle(
     Object.values(SECTIONS_STORAGE).reduce((acc, k) => {
@@ -44,8 +50,33 @@ const RemoteStorageDetails = ({
   );
 
   const renderActionMenu = useCallback(() => (
-    <MenuSection id="storage-details-actions" />
-  ), []);
+    <MenuSection id="storage-details-actions">
+      <Button
+        id="clickable-edit-storage"
+        buttonStyle="dropdownItem"
+        data-test-button-edit-storage
+        to={{
+          pathname: `${STORAGES_LIST_ROUTE}/edit/${id}`,
+        }}
+      >
+        <Icon size="small" icon="edit">
+          {intl.formatMessage({ id: 'ui-remote-storage.edit' })}
+        </Icon>
+      </Button>
+
+      <Button
+        id="clickable-delete-storage"
+        buttonStyle="dropdownItem"
+        data-test-button-delete-storage
+        onClick={onRemovestorage}
+      >
+        <Icon size="small" icon="trash">
+          {intl.formatMessage({ id: 'ui-remote-storage.delete' })}
+        </Icon>
+      </Button>
+    </MenuSection>
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  ), [id]);
 
   const closePane = useCallback(
     () => {
@@ -82,46 +113,53 @@ const RemoteStorageDetails = ({
         onToggle={toggleSection}
       >
         <Accordion
-          label={<FormattedMessage id="ui-remote-storage.details.title" />}
+          label={intl.formatMessage({ id: 'ui-remote-storage.details.title' })}
           id={SECTIONS_STORAGE.INFORMATION}
         >
           {storage.metadata && <ViewMetaData metadata={storage.metadata} />}
           <Row>
             <Col xs={6}>
-              <KeyValue label={<FormattedMessage id="ui-remote-storage.details.storageName" />}>
+              <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.details.storageName' })}>
                 {storage.name}
               </KeyValue>
             </Col>
             <Col xs={4}>
-              <KeyValue label={<FormattedMessage id="ui-remote-storage.details.providerName" />}>
-                {storage.providerName}
+              <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.details.providerName' })}>
+                {intl.formatMessage({ id: `ui-remote-storage.name.${storage.providerName}` })}
               </KeyValue>
             </Col>
             <Col xs={6}>
-              <KeyValue label={<FormattedMessage id="ui-remote-storage.details.url" />}>
+              <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.details.url' })}>
                 {storage.url}
               </KeyValue>
             </Col>
+            {storage.statusUrl && (
+              <Col xs={6}>
+                <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.details.statusUrl' })}>
+                  {storage.statusUrl}
+                </KeyValue>
+              </Col>
+            )}
           </Row>
         </Accordion>
 
         <Accordion
-          label={<FormattedMessage id="ui-remote-storage.accession.title" />}
+          label={intl.formatMessage({ id: 'ui-remote-storage.accession.title' })}
           id={SECTIONS_STORAGE.ACCESSION}
         >
-          <KeyValue label={<FormattedMessage id="ui-remote-storage.accession.schedule.title" />}>
-            <FormattedMessage
-              id="ui-remote-storage.accession.schedule.info"
-              values={{
+          <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.accession.schedule.title' })}>
+            {intl.formatMessage(
+              { id: 'ui-remote-storage.accession.schedule.info' },
+              {
                 delay: storage.accessionDelay,
                 unit: storage.accessionTimeUnit,
-              }}
-            />
+              },
+            )}
           </KeyValue>
         </Accordion>
 
         <Accordion
-          label={<FormattedMessage id="ui-remote-storage.retrieval.title" />}
+          label={intl.formatMessage({ id: 'ui-remote-storage.retrieval.title' })}
           id={SECTIONS_STORAGE.RETRIEVAL}
         />
       </AccordionSet>
@@ -133,6 +171,7 @@ RemoteStorageDetails.propTypes = {
   defaultWidth: PropTypes.string,
   storage: PropTypes.object,
   isLoading: PropTypes.bool,
+  onRemovestorage: PropTypes.func,
 };
 
 RemoteStorageDetails.defaultProps = {