+++ src/RemoteStorageSettings.js
@@ -7,20 +7,36 @@ import {
 } from '@folio/stripes/components';
 import RemoteStoragesListContainer from './RemoteStoragesList';
 import RemoteStorageDetails from './RemoteStorageDetails';
+import RemoteStorageEditorContainer from './RemoteStorageEditor';
+import CreateRomoteStoargeContainer from './CreateRemoteStorage';
 
 import { STORAGES_LIST_ROUTE } from './const';
 
 const RemoteStorageSettings = () => {
   return (
-    <Paneset
-      paneTitle={<FormattedMessage id="ui-remote-storage.meta.title" />}
-    >
-      <RemoteStoragesListContainer />
+    <>
+      <Paneset
+        paneTitle={<FormattedMessage id="ui-remote-storage.meta.title" />}
+      >
+        {/* <RemoteStoragesListContainer /> */}
+        <Route
+          path={STORAGES_LIST_ROUTE}
+          component={RemoteStoragesListContainer}
+        />
+        <Route
+          path={`${STORAGES_LIST_ROUTE}/view/:id`}
+          component={RemoteStorageDetails}
+        />
+      </Paneset>
       <Route
-        path={`${STORAGES_LIST_ROUTE}/view/:id`}
-        component={RemoteStorageDetails}
+        path={`${STORAGES_LIST_ROUTE}/edit/:id`}
+        component={RemoteStorageEditorContainer}
       />
-    </Paneset>
+      <Route
+        path={`${STORAGES_LIST_ROUTE}/create`}
+        component={CreateRomoteStoargeContainer}
+      />
+    </>
   );
 };