+++ src/RemoteStoragesList/RemoteStoragesListContainer.js
@@ -1,19 +1,30 @@
-import React, { useEffect, useState, useMemo } from 'react';
+import React, { useEffect, useState, useMemo, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import moment from 'moment';
+import { useLocation, useHistory } from 'react-router-dom';
+import { useIntl } from 'react-intl';
 
 import {
   stripesConnect,
   useStripes,
 } from '@folio/stripes/core';
 
+import { useLocationReset } from '@folio/stripes-acq-components';
+
 import RemoteStoragesList from './RemoteStoragesList';
 
+import { STORAGES_LIST_ROUTE } from '../const';
+
 const RemoteStoragesListContainer = ({
-  mutator,
+  mutator: originMutator,
 }) => {
   const stripes = useStripes();
+  const location = useLocation();
+  const history = useHistory();
+  const intl = useIntl();
 
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  const mutator = useMemo(() => originMutator, []);
   const [storagesList, setStoragesList] = useState([]);
   const [isLoading, setIsLoading] = useState();
   const [storagesCount, setStoragesCount] = useState();
@@ -22,7 +33,7 @@ const RemoteStoragesListContainer = ({
     moment.localeData(stripes.locale).longDateFormat('L')
   ), [stripes.locale]);
 
-  useEffect(() => {
+  const loadConfigurations = useCallback(() => {
     setIsLoading(true);
     mutator.configurations.GET().then(({ totalRecords, configurations }) => {
       setStoragesList(prev => [
@@ -32,19 +43,40 @@ const RemoteStoragesListContainer = ({
 
           return {
             ...storage,
+            providerName: intl.formatMessage({ id: `ui-remote-storage.name.${storage.providerName}` }),
             lastUpdate: lastUpdate.format(localeDateFormat),
           };
         }),
       ]);
       setStoragesCount(totalRecords);
     }).finally(() => setIsLoading(false));
-  }, [mutator, localeDateFormat]);
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [mutator.configurations, localeDateFormat]);
+
+  const refreshList = () => {
+    setStoragesList([]);
+    loadConfigurations();
+  };
+
+  useEffect(() => {
+    loadConfigurations();
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  const onCreateConfiguration = useCallback(() => {
+    history.push({
+      pathname: `${STORAGES_LIST_ROUTE}/create`,
+    });
+  }, [history]);
+
+  useLocationReset(history, location, STORAGES_LIST_ROUTE, refreshList);
 
   return (
     <RemoteStoragesList
       storages={storagesList}
       isLoading={isLoading}
       storagesCount={storagesCount}
+      onCreateConfiguration={onCreateConfiguration}
     />
   );
 };
@@ -55,11 +87,12 @@ RemoteStoragesListContainer.manifest = Object.freeze({
     path: 'remote-storage/configurations',
     accumulate: true,
     throwErrors: false,
+    clientGeneratePk: false,
+    fetch: false,
   },
 });
 
 RemoteStoragesListContainer.propTypes = {
-
   mutator: PropTypes.object.isRequired,
 };