+++ src/const.js
@@ -5,3 +5,9 @@ export const SECTIONS_STORAGE = {
   ACCESSION: 'accession',
   RETRIEVAL: 'retrieval',
 };
+
+export const TIME_UNITS = [
+  { label: 'second(s)', value: 'seconds' },
+  { label: 'minute(s)', value: 'minutes' },
+  { label: 'hour(s)', value: 'hours' },
+];