+++ CHANGELOG.md
@@ -5,4 +5,6 @@
 ### Stories
 * [UIRS-4](https://issues.folio.org/browse/UIRS-4) Project Setup: ui-remote-storage
 * [UIRS-3](https://issues.folio.org/browse/UIRS-3) Permission: Remote storage configuration
-* [UIRS-3](https://issues.folio.org/browse/UIRS-2) Remote storage configuration
+* [UIRS-2](https://issues.folio.org/browse/UIRS-2) Remote storage configuration
+* [UIRS-6](https://issues.folio.org/browse/UIRS-6) Edit remote storage configuration
+* [UIRS-7](https://issues.folio.org/browse/UIRS-7) Remove/delete remote storage configuration